Remove levels 0 and 1
Remove levels 0 and 1 because they are only testing levels.

Rename levels 2, 3, and 5 to be the new level 0, 1, and 2.

Remove Level 0 (Ian's Level) from the levelData list because it doesn't
fit with the theme of the other levels and needs polish.
